Move all config files to the same folder
