+++ .gitignore
@@ -2,6 +2,7 @@ node_modules/
 
 # dotenv environment variables files
 .env
+.env.*
 .env.*.local
 
 uploads/
