+++ src/middlewares/auth.middleware.js
@@ -1,15 +1,12 @@
 import jwt from "jsonwebtoken";
-import { JWT_SECRET } from "../../config/env.js";
+import { JWT_SECRET } from "../../configs/env.js";
 import User from "../models/user.model.js";
 
 const authorize = (req, res, next) => {
   try {
     let token;
 
-    if (
-      req.headers.authorization &&
-      req.headers.authorization.startsWith("Bearer")
-    ) {
+    if (req.headers.authorization && req.headers.authorization.startsWith("Bearer")) {
       token = req.headers.authorization.split(" ")[1];
     }
